Send request to ip address instead of fqdn
Fixes ZEN-22880

This change negates the need for DNS lookup in twisted when sending
a request to a Windows server.  It should both speed up requests
and allow for devices with no name resolution to be contacted.
Also added the --ipaddress,-s switch to winrm/winrs to enter
the ip address of the device.